move GenericLinkNode to own file, resolve raw host link to veth link
